Added environment variables for apcupsd.conf
A number of minor changes in this PR:

A start.sh script is now in place allowing for the addition of environment variables. A custom apcupsd.conf can be used through a host binding, or the environment variables will modify a container version of that file if a binding is not used. Files in the ./scripts directory are now conditionally moved in start.sh rather than the Dockerfile to support use of a persistent bound directory for these data files.

All files except for the README.md have been modified to remove <CR><LF> combos, and replace them with Linux standard <LF> only. A build.sh file has been added with the syntax for a buildx multi-arch build. This file is mostly for copy-and-paste use of the buildx command line, and should be modified as required for CPU architectures and target Docker Hub repositories and tags.